Update exercise 11.4

+++ i2p/exercise-11-4/main.c
@@ -107,7 +107,7 @@ void display_stats(char words[][MAX_LENGTH], int words_num) {
                 min_len = len;
         }
 
-        printf("Max length: %d\nMin length: %d\nAverage length: %f\n", max_len, min_len, (float) average_len / words_num);
+        printf("Max length: %d\nMin length: %d\nAverage length: %.2f\n", max_len, min_len, (float) average_len / words_num);
     }
     else
         printf("No words available\n");
@@ -143,8 +143,8 @@ void display_char_stats(char words[][MAX_LENGTH], int words_num) {
             current_occ = 0;
         }
 
-        printf("Occurrences: %d\nAverage occurrences/word: %d\nMax occurrences: %d (first in %s)\nMin occurrences: %d (first in %s)\n",
-            occurrences, occurrences / words_num, max_occ, words[max_idx], min_occ, words[min_idx]);
+        printf("Occurrences: %d\nAverage occurrences/word: %.2f\nMax occurrences: %d (first in %s)\nMin occurrences: %d (first in %s)\n",
+            occurrences, (float) occurrences / words_num, max_occ, words[max_idx], min_occ, words[min_idx]);
     }
     else
         printf("No words available\n");
